Cria interface para entidades, adiciona herança a classe entitie, adiciona entidade na classe commandResult, adiciona entidade para ser retornada nos resultados dos handlers das classes CreateUserHandler e UpdateUserHandler, altera o retorno dos mátodos da interface IUserApplicationService, adiciona metodo para validar se o body é nulo nos métodos dos Controllers da API e cria endpoints de update e delete na classe userController com suas respectivas rotas.

+++ GestaoDeUsuarios.WebApi/Controllers/UserController.cs
@@ -8,21 +8,27 @@ using GestaoDeUsuarios.Domain.Services;
 
 namespace GestaoDeUsuarios.WebApi.Controllers
 {
+    [RoutePrefix("api/user")]
     public class UserController : BaseController
     {
         private readonly IUserApplicationService _service;
 
         public UserController(IUserApplicationService userService) => _service = userService;
 
-        [HttpGet]
-        public Task<HttpResponseMessage> Get()
+        [Route("getall")]
+        [HttpPost]
+        public Task<HttpResponseMessage> GetAllUsers()
         {
             return CreateResponse(HttpStatusCode.Created, new { message = "API online" });
         }
 
+        [Route("create")]
         [HttpPost]
-        public Task<HttpResponseMessage> Post([FromBody]dynamic body)
+        public Task<HttpResponseMessage> CreateUser([FromBody]dynamic body)
         {
+            if(InvalidBody(body))
+                return CreateResponse(HttpStatusCode.BadRequest, "Dados inválidos"); 
+            
             var command = new CreateUserCommand(
                        nome: (string)body.nome,
                        sobrenome: (string)body.sobrenome,
@@ -31,12 +37,16 @@ namespace GestaoDeUsuarios.WebApi.Controllers
                    );
 
             var user = _service.Register(command);
-            return CreateResponse(HttpStatusCode.Created, command);
+            return CreateResponse(HttpStatusCode.Created, user.Entitie);
         }
 
+        [Route("update")]
         [HttpPost]
-        public Task<HttpResponseMessage> Put([FromBody]dynamic body)
+        public Task<HttpResponseMessage> UpdateUser([FromBody]dynamic body)
         {
+            if (InvalidBody(body))
+                return CreateResponse(HttpStatusCode.BadRequest, "Dados inválidos");
+
             var command = new UpdateUserCommand(
                        id: (Guid)body.id,
                        nome: (string)body.nome,
@@ -46,16 +56,21 @@ namespace GestaoDeUsuarios.WebApi.Controllers
                    );
 
             var user = _service.Update(command);
-            return CreateResponse(HttpStatusCode.OK, command);
+            return CreateResponse(HttpStatusCode.OK, user.Entitie);
         }
 
+        [Route("delete")]
         [HttpPost]
-        public Task<HttpResponseMessage> Delete([FromBody]dynamic body)
+        public Task<HttpResponseMessage> DeleteUser([FromBody]dynamic body)
         {
+            if (InvalidBody(body))
+                return CreateResponse(HttpStatusCode.BadRequest, "Dados inválidos");
+
             var command = new DeleteUserCommand((Guid)body.id);
 
             var user = _service.Delete(command);
-            return CreateResponse(HttpStatusCode.OK, command);
+
+            return CreateResponse(HttpStatusCode.OK, "");
         }
     }
 }
